feat: implement device register and updateInfo endpoints with validation
- Add RegisterDeviceRequest with 35+ field validation (nested screen, battery, admin, lockPattern objects)
- Add UpdateDeviceInfoRequest for lightweight 6-field updates
- Implement register() with camelCase-to-snake_case mapping, device upsert by device_uid + user_id, and device_details upsert
- Implement updateInfo() for brand, apiGrade, langCode, phoneNumber updates
- Fix lock_quality column type from smallInteger to integer (supports Android PASSWORD_QUALITY_* constants up to 327680)
- Update existing auth tests to work with new validation requirements
- Add 7 new tests: register validation, full registration, upsert, update validation, update success, update DB verification, update 404
- All 18 API tests passing (80 assertions)

+++ app/database/migrations/2026_04_10_000004_create_device_details_table.php
@@ -52,7 +52,7 @@ return new class extends Migration
             $table->boolean('is_keyguard_secure')->nullable();
             $table->boolean('is_device_secure')->nullable();
             $table->boolean('in_keyguard_restricted_input_mode')->nullable();
-            $table->smallInteger('lock_quality')->default(-1);
+            $table->integer('lock_quality')->default(-1);
 
             // Battery (BatteryLevelVO)
             $table->float('battery_percent')->nullable();
